requirements.txt
flask==2.3.3
flask-cors==4.0.0
groq==0.9.0
Pillow==10.0.1
pyPDF2==3.0.1